Resplit CommandBus interface
