+++ src/CommandBus.php
@@ -1,55 +1,20 @@
 <?php
-
 namespace League\Tactician;
 
-use League\Tactician\Exception\CommandWasNotHandledException;
-
 /**
- * Receives a command and modifies or dispatches it to a handler in some way
+ * Receives a command and sends it through a chain of middleware for processing.
+ *
+ * This interface is only useful for mocks and legacy decorators to implement,
+ * you should always use StandardCommandBus and implement Middleware in your
+ * application, rather than decorate the command bus for additional behavior.
  */
-class CommandBus
+interface CommandBus
 {
-    /**
-     * @var Closure
-     */
-    private $middlewareChain;
-
-    /**
-     * @param Middleware[] $middleware
-     */
-    public function __construct(array $middleware)
-    {
-        $this->middlewareChain = $this->createExecutionChain($middleware);
-    }
-
     /**
      * Executes the given command and optionally returns a value
      *
      * @param Command $command
      * @return mixed
      */
-    public function execute(Command $command)
-    {
-        $middlewareChain = $this->middlewareChain;
-        return $middlewareChain($command);
-    }
-
-    /**
-     * @param Middleware[] $middlewareList
-     * @return \Closure
-     */
-    protected function createExecutionChain($middlewareList)
-    {
-        $lastCallable = function (Command $command) {
-            // the final callable is a no-op
-        };
-
-        while ($middleware = array_pop($middlewareList)) {
-            $lastCallable = function (Command $command) use ($middleware, $lastCallable) {
-                return $middleware->execute($command, $lastCallable);
-            };
-        }
-
-        return $lastCallable;
-    }
+    public function execute(Command $command);
 }
