+++ tests/Setup/QuickStartTest.php
@@ -1,7 +1,7 @@
 <?php
 namespace League\Tactician\Tests\Setup;
 
-use League\Tactician\CommandBus;
+use League\Tactician\StandardCommandBus;
 use League\Tactician\Setup\QuickStart;
 use League\Tactician\Tests\Fixtures\Command\AddTaskCommand;
 use League\Tactician\Tests\Fixtures\Command\CompleteTaskCommand;
@@ -12,7 +12,7 @@ class QuickStartTest extends \PHPUnit_Framework_TestCase
     public function testReturnsACommandBus()
     {
         $commandBus = QuickStart::create([]);
-        $this->assertInstanceOf(CommandBus::class, $commandBus);
+        $this->assertInstanceOf(StandardCommandBus::class, $commandBus);
     }
 
     public function testCommandToHandlerMapIsProperlyConfigured()
